feat(VerticalBarChart): create interfaces for chart data and component props to enhance type safety and organization
